pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "fastapi>=0.115.11",
    "openai>=1.65.5",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "snowflake-connector-python>=3.14.0",
    "sqlalchemy>=2.0.38",
    "streamlit>=1.43.1",
    "trafilatura>=2.0.0",
    "uvicorn>=0.34.0",
]
